fnt: Remove FntChars
As far as I can tell, it only tells us about the number of characters
that will get defined, but we can easily just check for that number.
Keeping this around seems unnecessary.
If it turns out that `chars` can hold more information, we might add it
again in the future.